% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_categories_content.R
\name{check_categories_content}
\alias{check_categories_content}
\title{Check Category Content}
\usage{
check_categories_content(dc)
}
\arguments{
\item{dc}{output from the compare_datapoints function}
}
\description{
Compares the rows of categories in two datapoints
}
\keyword{QC}
\keyword{dataset,}
\keyword{utility,}

