% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_datapoints_categories.R
\name{check_datapoints_categories}
\alias{check_datapoints_categories}
\title{Check Data Points Categories}
\usage{
check_datapoints_categories(datapoints.compare)
}
\arguments{
\item{f}{a data points object provided by the fulfiller}

\item{q}{a data points object produced by the qc'er}
}
\description{
Returns an object that holds two datasets in data points format, use
with functions that compare datapoints
}
\keyword{QC}
\keyword{dataset,}
\keyword{utility,}

