% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_value_content.R
\name{check_value_content}
\alias{check_value_content}
\title{Check Value Content}
\usage{
check_value_content(dc)
}
\arguments{
\item{dc}{output from the compare_datapoints function}
}
\description{
Compares the rows of values in two datapoints
}
\keyword{QC}
\keyword{dataset,}
\keyword{utility,}

