% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean}
\alias{clean}
\title{Clean Dataset}
\usage{
clean(ds)
}
\arguments{
\item{ds}{The source dataset}
}
\description{
Transforms dataset by rounding numerics and removing all
variations of
NULL and NA and replacing them with blank characters. This can be used
to quickly test datasets with weird formats.
}
\details{
Clean a dataset
}
\keyword{QC}
\keyword{dataset,}
\keyword{utility,}
