% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_datapoints.R
\name{compare_datapoints}
\alias{compare_datapoints}
\title{Compare Data Points}
\usage{
compare_datapoints(f, q)
}
\arguments{
\item{f}{a data points object provided by the fulfiller}

\item{q}{a data points object produced by the qc'er}
}
\description{
Returns an object that holds two datasets in data points format, use
with functions that compare datapoints
}
\examples{

USE WIDE WITH DATAPOINTS

datapoints(quakes, c(1,2,5)) \%>\%
wide() \%>\%
head()

}
\keyword{QC}
\keyword{dataset,}
\keyword{utility,}

