% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_points.R
\name{data_points}
\alias{data_points}
\title{Contains Both Datasets}
\usage{
data_points(a, b)
}
\arguments{
\item{a}{a data points object provided by the fulfiller in the first dataset}

\item{b}{a data points object produced by the qc analyst in the second dataset}
}
\description{
Returns an object that holds two datasets in data points format, use
with functions that compare datapoints
}
