% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPoints.R
\name{DataPoints}
\alias{DataPoints}
\title{Data Points}
\usage{
DataPoints(df, category.cols = NULL, cleaning.function = NULL)
}
\arguments{
\item{df}{The source dataset}
}
\description{
This function makes it easier to look at datasets
}
\examples{
d <- DataPoints(quakes, c(1,2,5))
}
\keyword{QC}
\keyword{dataset,}
\keyword{utility,}

