% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter}
\alias{filter}
\title{Filter}
\usage{
filter(d, ...)
}
\arguments{
\item{d}{data points, compare_datapoints or dataframe object}
}
\description{
Filters data points
}
\keyword{QC}
\keyword{dataset,}
\keyword{utility,}

