% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_sql.R
\name{filter_sql}
\alias{filter_sql}
\title{SQL Filter}
\usage{
filter_sql(d, filter)
}
\arguments{
\item{d}{data points, compare_datapoints or dataframe object}
}
\description{
Filters data points
}
\examples{

USE WIDE WITH DATAPOINTS

datapoints(quakes, c(1,2,5)) \%>\%
filter_sql() \%>\%
head()

}
\keyword{QC}
\keyword{dataset,}
\keyword{utility,}

