% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_matching_report.R
\name{html_matching_report}
\alias{html_matching_report}
\title{Present Results of a Match}
\usage{
html_matching_report(dp, qc_title = "")
}
\arguments{
\item{dp}{is a compare_datapoints objects containing both the f and q datapoints}

\item{qc_title}{the title that you would like to appear in the QC match report}
}
\description{
Returns a HTML formated report detailing the results of a QC Match
Analysis
}
\keyword{QC}

