% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.data.points.R
\name{is.data.points}
\alias{is.data.points}
\title{Is Data Frame in Data Points Format?}
\usage{
is.data.points(df)
}
\arguments{
\item{df}{The source dataset}
}
\description{
Check to see if the last two columns are always a named variable and value.
variable is a character type and would be a column header in a tidy dataframe. The preceding
columns are all categories that you can select from.
}
\keyword{QC,}
\keyword{matching}

