% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.long.R
\name{is.long}
\alias{is.long}
\title{Is Dataframe In Long Format?}
\usage{
is.long(df)
}
\arguments{
\item{df}{The source dataset}
}
\description{
Make sure the dataset is in long format and that the last two columns are always a named variable and value.
}
