% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_category_content.R
\name{match_category_content}
\alias{match_category_content}
\title{Match Category Content}
\usage{
match_category_content(dp)
}
\arguments{
\item{dp}{a datapoints object with f and q datasets}
}
\description{
Compares the rows of categories in two datapoints
}
\keyword{QC,}
\keyword{matching}

