% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_category_metadata.R
\name{match_category_metadata}
\alias{match_category_metadata}
\title{Match Category Meta Data}
\usage{
match_category_metadata(dp)
}
\arguments{
\item{dp}{output from the compare_datapoints function}
}
\description{
Compares the types of categories in datapoints
}
\keyword{QC,}
\keyword{matching}

