% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_data_points.R
\name{match_data_points}
\alias{match_data_points}
\title{Match All Data Points}
\usage{
match_data_points(dp)
}
\arguments{
\item{dp}{\link[datapointsr]{data_points} object containing the fulfiller dataset f and the QC dataset q}
}
\description{
match_data_points checks all content including column metadata. If the match is not 100% equal then match_data_points will return datasets describing the differences between the datasets.
}
\details{
Attempts to do an exact 1:1 match between two datasets stored in a data points object.
}

