% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_values.R
\name{match_values}
\alias{match_values}
\title{Check All Content}
\usage{
match_values(dp)
}
\arguments{
\item{dp}{output from the compare_datapoints function}
}
\description{
Compares the rows of values in two datapoints
}
\keyword{QC}
\keyword{dataset,}
\keyword{utility,}

