% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/present_qc_match_report.R
\name{present_qc_match_report}
\alias{present_qc_match_report}
\title{Present QC Report}
\usage{
present_qc_match_report(dc, qc_title = "")
}
\arguments{
\item{dc}{is a compare_datapoints objects containing both the f and q datapoints}

\item{qc_title}{the title that you would like to appear in the QC match report}
}
\description{
Returns a HTML formated report detailing the results of a QC Match
Analysis
}
\keyword{QC}

