% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_match_values.R
\name{report_match_values}
\alias{report_match_values}
\title{Present Results of a Value Match}
\usage{
report_match_values(dp, qc_title = "")
}
\arguments{
\item{dp}{is a compare_datapoints objects containing both the f and q datapoints}

\item{qc_title}{the title that you would like to appear in the QC match report}
}
\description{
Returns a HTML formated report detailing the results of a QC Match
Analysis
}
\keyword{QC}

