% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_a.R
\name{show_a}
\alias{show_a}
\title{Show F Mismatched Data Points}
\usage{
show_a(match_dp)
}
\arguments{
\item{match_dp}{match_data_points object populated with matching information.}
}
\description{
pulls out a dataframe that shows only values that appeared in the 1st(a) dataset
}
