% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_as_html.R
\name{show_as_html}
\alias{show_as_html}
\title{Show Match as HTML}
\usage{
show_as_html(matched_data_points, title)
}
\arguments{
\item{matched_data_points}{\link[datapointsr]{matched_data_points} analysis object}

\item{title}{text title to include in the report}
}
\description{
Summarizes the results of a match QC with HTML for R Markdown display
}
