% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_b.R
\name{show_b}
\alias{show_b}
\title{Show B Mismatched Data Points}
\usage{
show_b(match_dp)
}
\arguments{
\item{match_dp}{match_data_points object populated with matching information.}
}
\description{
pulls out a dataframe that shows only values that appeared in the 2nd(b) dataset
}
