% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_both.R
\name{show_both}
\alias{show_both}
\title{Show Mismatched Data Points}
\usage{
show_both(match_dp)
}
\arguments{
\item{match_dp}{match_data_points object populated with matching information.}
}
\description{
pulls out a dataframe that shows only values that are in both datasets but are mismatched
}
\details{
Look at mismatched data points.
}
