% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_categories.R
\name{show_categories}
\alias{show_categories}
\title{Match Category Content}
\usage{
show_categories(dp)
}
\arguments{
\item{dp}{a datapoints object with f and q datasets}
}
\description{
Compares the rows of categories in two datapoints
}
\keyword{QC,}
\keyword{matching}

