% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_category_metadata.R
\name{show_category_metadata}
\alias{show_category_metadata}
\title{Show Category Meta Data}
\usage{
show_category_metadata(dp)
}
\arguments{
\item{dp}{output from the compare_datapoints function}
}
\description{
Shows the types of categories in datapoints
}
\keyword{QC,}
\keyword{matching}

