% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_f.R
\name{show_f}
\alias{show_f}
\title{Show F Mismatched Data Points}
\usage{
show_f(match_dp)
}
\arguments{
\item{match_dp}{match_data_points object populated with matching information.}
}
\description{
pulls out a dataframe that shows only values that appeared in the fulfiller dataset
}
