% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_in_console.R
\name{show_in_console}
\alias{show_in_console}
\title{Show Match in Console}
\usage{
show_in_console(matched_data_points)
}
\arguments{
\item{matched_data_points}{match_dp match_data_points object populated with matching information.}
}
\description{
Summarizes the results of a match QC
}
