% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_q.R
\name{show_q}
\alias{show_q}
\title{Show Q Mismatched Data Points}
\usage{
show_q(match_dp)
}
\arguments{
\item{match_dp}{match_data_points object populated with matching information.}
}
\description{
pulls out a dataframe that shows only values that appeared in the qc'er dataset
}
