% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_values.R
\name{show_values}
\alias{show_values}
\title{Show Values}
\usage{
show_values(dp)
}
\arguments{
\item{dp}{data.points object}
}
\description{
Shows the values in two datapoints
}
\keyword{QC,}
\keyword{matching}

