% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_variables.R
\name{show_variables}
\alias{show_variables}
\title{Show Variables}
\usage{
show_variables(dp)
}
\arguments{
\item{dp}{data.points object}
}
\description{
Shows the variables in two datapoints
}
\keyword{QC,}
\keyword{matching}

