% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wide.R
\name{wide}
\alias{wide}
\title{Wide Data Points Format}
\usage{
wide(df)
}
\arguments{
\item{df}{a \link[datapointsr]{long} format dataframe}
}
\description{
Returns a \link[datapointsr]{long} dataset as wide format
}
