% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_categories_meta.R
\name{check_categories_meta}
\alias{check_categories_meta}
\title{Check Meta Categories}
\usage{
check_categories_meta(dc)
}
\arguments{
\item{dc}{output from the compare_datapoints function}
}
\description{
Compares the types of categories in datapoints
}
\keyword{QC}
\keyword{dataset,}
\keyword{utility,}

