% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_datapoints.R
\name{compare_datapoints}
\alias{compare_datapoints}
\title{Compare Data Points}
\usage{
compare_datapoints(f, q)
}
\arguments{
\item{f}{a data points object provided by the fulfiller}

\item{q}{a data points object produced by the qc'er}
}
\description{
Returns an object that holds two datasets in data points format, use
with functions that compare datapoints
}
\examples{
f <- quakes
q <- quakes

q <- q[order(q$long), ]
q[56, 3] <- 1234

head(f)
head(q)

f <- datapoints(f, c(1,2,5))
q <- datapoints(q, c(1,2,5))

head(f)
head(q)

d <- compare_datapoints(f, q)
}
\keyword{QC}
\keyword{dataset,}
\keyword{utility,}

