% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_points.R
\name{data_points}
\alias{data_points}
\title{Contains Both Datasets}
\usage{
data_points(f, q)
}
\arguments{
\item{f}{a data points object provided by the fulfiller}

\item{q}{a data points object produced by the qc analyst}
}
\description{
Returns an object that holds two datasets in data points format, use
with functions that compare datapoints
}

