% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPoints.R
\name{datapoints}
\alias{datapoints}
\title{Data Points}
\usage{
datapoints(df, category.cols)
}
\arguments{
\item{df}{The source dataset}

\item{category.cols}{The columns used to filter the dataset}
}
\description{
This object makes it easier to look at datasets
}
\examples{

BUILD DATAPOINTS OBJECT FOR THE BUILT IN QUAKES DATASET

d <- datapoints(quakes, c(1,2,5))

}
\keyword{QC}
\keyword{dataset,}
\keyword{utility,}

