% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wide.R
\name{wide}
\alias{wide}
\title{Wide}
\usage{
wide(dp)
}
\arguments{
\item{dp}{a data points object}
}
\description{
Returns a dataset in a wide format
}
\examples{

USE WIDE WITH DATAPOINTS

datapoints(quakes, c(1,2,5)) \%>\%
wide() \%>\%
head()


}
\keyword{QC}
\keyword{dataset,}
\keyword{utility,}

